+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `heteroZip` and `heteroZipWith`.
 * Add `traverse_` and `for_`.
+* Add `nubOrd` and `nubOrdBy`.
 
 # 0.1.1